Read CA cert from service acct, check server readiness from endpoint
